Extract ConsumerConfig into own class
This matches the PEP8 convention, whilst splitting the changes into two
commits to allows future readers to see where the refactor and move
happened in two steps
