+++ OpenStack-Rabbit-Consumer/rabbit_consumer/openstack_api.py
@@ -4,7 +4,8 @@ import requests
 from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
 
-from rabbit_consumer.rabbit_consumer import RabbitConsumer, ConsumerConfig
+from rabbit_consumer.rabbit_consumer import RabbitConsumer
+from rabbit_consumer.consumer_config import ConsumerConfig
 
 logger = logging.getLogger(__name__)
